% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d2015}
\alias{d2015}
\title{d2015}
\format{
coursera datasets...
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
\url{http://www.coursera.com/}
}
\usage{
d2015
}
\description{
readed dataset accident_2015.csv.bz2
}
\keyword{datasets}
