% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{numeric value of the state (USA)}

\item{year}{year to get data retrieved}
}
\value{
plot with the state and coordinates of the data
}
\description{
This function return a plot with the state limits and the coordinates of the cases within it
}
\examples{
\dontrun{
fars_map_state(1, 2013)
}
}
