% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{File name or locations with the data}
}
\value{
tibble with the readed data
}
\description{
Function to read data from data from the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System, which is a nationwide
census providing the American public yearly data regarding fatal injuries
suffered in motor vehicle traffic crashes
}
\examples{
\dontrun{
fars_read('./data-raw/accident_2013.csv.bz2')
}
}
