% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{years to find data}
}
\value{
List of tibbles with two fields: MONTH and year. If year not in data
a warning message will be printed
}
\description{
This function returns a list of the tibbles from years founded in data.
}
\details{
Data will be readed from the execution place of the file
The data in csv.bz2 format need to be in the working directory
}
\examples{
\dontrun{
x <- fars_read_years(c(2000,2013,2015, 2021))
x[[2]]$MONTH
}
}
