% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year of the data}
}
\value{
character with the name of the file (with extenion .csv.bz2)
}
\description{
Funtion to stablish a normalized filename given a year number
}
\examples{
\dontrun{
make_filename(2000)
}
}
